appointment-service/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.22
alembic==1.12.0
asyncpg==0.28.0
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.0

# Date/time handling for appointments
pytz==2023.3.post1
icalendar==5.0.7

# Validation
email-validator==2.0.0.post2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Notifications
jinja2==3.1.2
aiosmtplib==2.0.2

# Google Calendar integration
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Microsoft Calendar integration
msal==1.24.1
